refactor(run-protocol): factor applyDelta etc. out of vault.js
